+++ telegram-bot/src/nativeMain/kotlin/eu/vendeli/tgbot/implementations/ClassManagerImpl.native.kt
@@ -1,5 +1,6 @@
 package eu.vendeli.tgbot.implementations
 
+import eu.vendeli.tgbot.TelegramBot
 import eu.vendeli.tgbot.interfaces.ClassManager
 import kotlin.reflect.KClass
 
@@ -9,8 +10,13 @@ import kotlin.reflect.KClass
  * @constructor Create empty ClassManagerImpl
  */
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-actual class ClassManagerImpl actual constructor() : ClassManager {
-    override fun getInstance(kClass: KClass<*>, vararg initParams: Any?): Any {
-        TODO("Not yet implemented")
-    }
+actual class ClassManagerImpl : ClassManager {
+    @Suppress("PropertyName")
+    val _INSTANCES: MutableMap<KClass<*>, Any> = mutableMapOf()
+    override fun getInstance(kClass: KClass<*>, vararg initParams: Any?): Any = _INSTANCES[kClass]
+        ?: error("Class $kClass not found.")
+}
+
+fun TelegramBot.addInstances(block: MutableMap<KClass<*>, Any>.() -> Unit) {
+    (config.classManager as? ClassManagerImpl)?._INSTANCES?.apply(block)
 }